move  all of ams
